Add code that checks which page we're on by checking the title element. This is used to run specific functions. This commit also adds code that checks the score at the end of a game against the highscore. Finally, scores are stored in localstorage and written to the DOM when the player browses to the high score page.
This is quite an extensive commit. First off, the function that runs on page load has been changed. It now checks the title element of the page, and uses that to run certain functions. On the home page it runs functions to run the game. On the highscores pages, it runs a function to display the highscores in localstorage. On the game settings page, it currently runs no function. I've also added a function that takes the score at the end of a game and compares it to the score in localstorage. Currently there is just one score being stored in localstorage, but I will extend this to a total of 10 scores.