Add difficulty settings back to the game.
This commit adds difficulty buttons to the Game settings page. Which difficulty has been choses is saved in localStorage. This localStorage is also read to display which difficulty had been chosen previously, so the game remembers which difficulty you chose last. The difficulty setting is used to determine the starting length of the pattern: easy = 1, medium = 5 and hard = 9. This only applies to the first run: after that the length of the previous sequence is used to determine pattern length.