Update readme.md with results from first code validation and fix some quick errors.
I've run the code validators for HTML and CSS and updated the testing section in readme.md. I've also installed JSHint. The first run produced a lot of errors, but this was mainly due to JSHint not being configured to use ECMAScript 6 settings. After configuring this, the amount of errors dropped considerably. Most of the remaining errors were missing semicolons. This commit adds those.