Change the formatting of the code and add docstrings to functions.
In this commit I've run Prettier to change the formatting of the code to make it more legible and standardized. I've also added docstrings to the functions, so that when you hover over functions in most modern code editors, you will see this docstring. This was a tip from my mentor Jack Wachira.