Merge the two for loops in the saveScore function.
There were two for loops in the saveScore function, one loop to save all the values to an array, the other to check all the values vs the endscore. These two have now been combined into one loop, that writes all the values to an array and checks all the values vs the endscore at the same time.